requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-heroku==0.3.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.2.1
python-dotenv==1.0.1
service-identity==24.2.0
setuptools==78.1.0
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
whitenoise==6.8.2
zope.interface==7.1.1
